backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pandas==2.2.2
openpyxl==3.1.5
orjson==3.10.7
numpy==2.0.1
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-dotenv==1.0.1
psycopg[binary]==3.2.9
pyodbc==5.2.0




